requirements.txt
python-telegram-bot
PySocks
future
certifi